fortune-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==0.28.1
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.2
